+++ lib/public/components/NavBar/index.js
@@ -19,11 +19,7 @@ import { absoluteFrontLink } from '../common/navigation/absoluteFrontLink.js';
 import { externalLinks } from '../common/navigation/externalLinks.js';
 import { configurationService } from '../../services/configurationService.js';
 import { selectionDropdown } from '../common/selection/dropdown/selectionDropdown.js';
-import { SelectionDropdownModel } from '../common/selection/dropdown/SelectionDropdownModel.js';
-import { getDetectorExpertRoleName } from '../../utilities/getDetectorExpertRoleName.js';
-import { dplDetectorsProvider } from '../../services/detectors/dplDetectorsProvider.js';
-
-let isChecked = true;
+import { BkpRolesSelectionModel } from '../../models/BkpRolesSelectionModel.js';
 
 /**
  * Generic tab button for website navigation purposes
@@ -114,46 +110,6 @@ const dropdownSubMenu = (dropdownButtonContent, isOpened, onclick, entries, { me
         h('.dropdown-menu', menuAttributes, entries),
     ]);
 
-/**
- * Model storing state of selected Bkp roles (for TEST only)
- */
-export class BkpRolesSelectionModel extends SelectionDropdownModel {
-    // eslint-disable-next-line valid-jsdoc
-    /**
-     * @inheritDoc
-     */
-    constructor() {
-        super({ defaultSelection: [{ value: 'admin' }] });
-    }
-
-    // eslint-disable-next-line valid-jsdoc
-    /**
-     * @inheritDoc
-     */
-    _initialize() {
-        /**
-         * Update dropdown model options with DPL detector provider remote data
-         * @return {void}
-         */
-        const update = () => this.setAvailableOptions(dplDetectorsProvider.physical$.getCurrent().match({
-            Success: (detectors) => [
-                'admin',
-                ...detectors.map(({ name }) => getDetectorExpertRoleName(name)),
-            ]
-                .map((roleName) => ({
-                    value: roleName,
-                    rawLabel: roleName,
-                })),
-            Other: () => [{ value: 'admin', rawLabel: 'admin' }],
-        }));
-
-        setTimeout(() => {
-            dplDetectorsProvider.physical$.observe(update);
-            update();
-        }, 1);
-    }
-}
-
 const bkpRolesSelectionModel = new BkpRolesSelectionModel();
 
 /**
@@ -172,12 +128,7 @@ function navBar(model) {
 
     if (model.session.personid === 0) { // Anonymous user has id 0
         bkpRolesSelectionModel.observe(() => {
-            const options = bkpRolesSelectionModel.options.map(({ value }) => value);
-            model.session.access = [
-                ...(model.session.access ?? [])
-                    .filter((roleName) => !options.includes(roleName)),
-                ...bkpRolesSelectionModel.selected,
-            ];
+            model.session.access = bkpRolesSelectionModel.selected;
         });
     }
